Adds gettext-compile/extract as yarn tasks, ensures nominal behavior
* yarn gettext-extract/compile will  supplants gup task  
* yarn gettext-compile will also run yarn available-languages(generates available-languages.json)
* commit also includes the new condensed manageiq-ui-service.json file (instead of seperate folders for each translated language)
* manageiq-ui-service.pot has been updated

All translation behavior is nominal after this commit.
